requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements/requirements.txt requirements/requirements.in
#
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via requests
idna==3.3
    # via requests
pycryptodome==3.15.0
    # via -r requirements/requirements.in
python-dotenv==0.20.0
    # via -r requirements/requirements.in
requests==2.28.1
    # via -r requirements/requirements.in
urllib3==1.26.11
    # via requests
